+++ src/sdk/PnP.Core/Model/SharePoint/Pages/Internal/Page.cs
@@ -1655,7 +1655,7 @@ namespace PnP.Core.Model.SharePoint
                 currentSection.Collapsible = zoneGroupMetadata.Type == 1;
                 currentSection.SectionType = zoneGroupMetadata.Type;
                 currentSection.DisplayName = zoneGroupMetadata.DisplayName;
-                currentSection.HeadingLevel = zoneGroupMetadata.HeadingLevel;
+                currentSection.HeadingLevel = zoneGroupMetadata.HeadingLevel != 0 ? zoneGroupMetadata.HeadingLevel : (int)CollapsibleSectionHeadingLevel.Heading2;
                 currentSection.IsExpanded = zoneGroupMetadata.IsExpanded;
                 currentSection.ShowDividerLine = zoneGroupMetadata.ShowDividerLine;
                 if (zoneGroupMetadata.IconAlignment != null)