+++ src/sdk/CHANGELOG.md
@@ -15,6 +15,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 - Page API: support for flexible sections! #1682 [czullu - Christian Zuellig]
 - Page API: fix to retain empty, collapsible sections #1682 [czullu - Christian Zuellig]
 - Page API: fix to retain NoImage setting on custom page header #1682 [czullu - Christian Zuellig]
+- Page API: support for setting the header size of collapsible sections #1687 [shunsaker - Shayne Hunsaker]
 
 ### Changed